wdt: stop watchdog after panic [1/1]
PD#SWPL-143567

Problem:
When debugging, there will be a lot of printing to be
output after the panic, so it is very time-consuming.
In order to prevent the watchdog from causing a reset,
we stop it after panic.

Solution:
Add functions and provide two ways to enable:
1. Add wdt_stop_after_panic=1 to bootargs.
2. echo 1 > /sys/module/amlogic_watchdog/parameters/stop_after_panic

Verify:
T3/T982

Change-Id: I3f54f38f3645b2fb4fdfbbc6016e6f644b91eea3
Signed-off-by: Huqiang Qin <huqiang.qin@amlogic.com>