audio: Fix the problem that a4 acodec cannot record sound [1/1]
PD#SWPL-170330

Problem:
When the acodec driver is loaded, the kernel cannot read the acodec_1
register value, thereby misjudging the connection status of the audio
path, causing the captrue software input path to be disconnected and
unable to record data.

Solution:
Add the operation of setting the acodec register to the
default value at probe.

Verify:
A113L2-BA400

Change-Id: Ic3a313b652cbd47ad41aba34fe6a7183dda8d211
Signed-off-by: shu.wang <shu.wang@amlogic.com>