hdmitx21: Optimization of standby power consumption [1/1]
PD#SWPL-172001

Problem:
When s7d is in standby and wakes up, it will turn off and on the vpu
power domain. When it is turned off, the reg of the relevant modules
will be reset, the output no signal when it wakes up.

Solution:
hdmitx driver needs to reinitialize the required top register when
it wake up

Verify:
s7d

Test:
DRM-TX-78, DRM-TX-79, DRM-TX-80, DRM-TX-81

Change-Id: I0414be14d4ee28b8493b61122201b6b2c54569a0
Signed-off-by: zhou.han <zhou.han@amlogic.com>