crypto: disable link mode when size larger than 2^17 [1/1]
PD#SWPL-181936

Problem:
When input size is large and link mode is enabled, we failed to get
corret result.
The reason is if link mode is enabled, the buffers are scattered
around and the size could be either large or small.
We have to use byte mode to move the data.
However, the maximum supported length for byte mode is 2^17.
If over, the data will not be processed.

Solution:
Disable link mode if data size overflow happens.

Verify:
S6

Change-Id: I22d5d09d2a3f256f769303749364ddbeb46a5f89
Signed-off-by: Matthew Shyu <matthew.shyu@amlogic.com>