spinand: fixup reserve dtb size is different with bl33 [1/1]
PD#SWPL-177461

Problem:
bl33 crash after reboot in kernel, this is due to the dtb lose.
the reason for the loss of dtb is dtb size in kernel is different with bl33
when flash spi nand page size is 4096.

Solution:
get rsv size from dtb, which had written by bl33

Verify:
C308L-AW419

Change-Id: I187d8db178d28e39c8f5085b1720033f211370f3
Signed-off-by: zhikui.cui <zhikui.cui@amlogic.com>