ramdump: fix buildin err for parse info. [1/1]
PD#SWPL-191355

Problem:
For the low_mem project, the ramdump driver is
built-in. If many kernel configs are removed
during kernel size trimming, it may lead to build
errors in ramdump parsing info.
The removed configs info can be found in CL 486868.

Solution:
Romove parse info of kimage_voffset when ramdump
drv is built-in.

Verify:
sc2

Change-Id: I36b533bd06e56ca66b87850d01c0b0ad57c3a73c
Signed-off-by: dongqing.li <dongqing.li@amlogic.com>