drm: optimize frac_rate_policy flow [1/1]
PD#SWPL-189168

Problem:
BDS AndroidU: bootup without hdmi, and connected to HDMI after startup,
the output is not full screen.

Android-U:
1 The value of frac_rate_policy env is 1 in uboot, and vout outputs cvbs
2 No HDMI cable is connected during the logo_init.
meson_hdmitx_encoder_atomic_check does not go in, and the value of property
is still 0 by default.
3 TX has been successfully configured with 2160p, hwc only sets the frac
video mode and property is 0. hwc check fails, and 1080p is displayed
OTT does not have this problem, because hwc mode list has integer
+ decimal, so can match.

Android-T
Steps 1 and 2 are the same as U,and hwc read node on T. After plugin,
the default value of 0 for property will be updated to the node,
and then step3 above will happen again.

Solution:
1 Get the value directly from the node
2 Delete the process that copies values from property to nodes

Verify:
T7C

Test:
DRM-TX-20

Change-Id: I57c59e3ee013d58bd05c1679eb4b4f2e2592ef90
Signed-off-by: zhou.han <zhou.han@amlogic.com>