ramdump: memcpy 0-24MB on panic for ker-32bit [2/2]
PD#SWPL-192100

Problem:
The 32-bit kernel cannot use the ramdump function.
Because the .text section is typically at 0x00108000,
which will be overwritten by the BL2E/BL33 after
reboot. And the PC side is unable to parse it.

Solution:
1. After kernel panic, use DMA to copy the 0-24MB
region to the 0x09000000 area.
2. Reboot to bl33z, fill compress_section[1] with
the 0x09000000 region.
3. Analyze it using the same method as before.

Verify:
sc2

Change-Id: I3e22beb06321289c6643d3f05ec2d910d5ce0beb
Signed-off-by: dongqing.li <dongqing.li@amlogic.com>