host: add efuse register for dsp enable check [1/1]
PD#SWPL-187343

Problem:
Some chips have their dsp disabled by efuse, which would cause fail
starting of dsp. In this case, the reserved resources in the dts
should be released.

Solution:
Read the efuse register to confirm whether dsp is disabled or not
and release the reserved space in dts when it is disabled.

Verify:
sc2/s6

Change-Id: If9ebb4c26d72fe225354c23063b96efa73a2de7d
Signed-off-by: Liming Xue <liming.xue@amlogic.com>