clk: Fix the SCMI shmem cache issue [2/2]
PD#SWPL-196548
PD#SWPL-196423

Problem:
In the Kernel, reading and writing SCMI shmem is non-cache, but in bl31
it is cache. Therefore, it will cause an abnormal SCMI communication
between BL31 and the Kernel due to the cache synchronization issue.

Solution:
SCMI shmem of bl31 is changed to non-cache.

Verify:
A113X2/S905C2

Change-Id: I911a9dd3382b195cb7a16719fa13e290446ba160
Signed-off-by: Chuan Liu <chuan.liu@amlogic.com>