clk: Fix the SCMI shmem cache issue [3/3]
PD#SWPL-196548

Problem:
In the Kernel, reading and writing SCMI shmem is non-cache, but in bl31
it is cache. Therefore, it will cause an abnormal SCMI communication
between BL31 and the Kernel due to the cache synchronization issue.

Solution:
SCMI shmem of bl31 is changed to non-cache.

Verify:
A113L2

Change-Id: Ib113b59209169b57e224f98e0b180b0322764cdf
Signed-off-by: Chuan Liu <chuan.liu@amlogic.com>