spisg: fix transmit incomplete issue when trig stop [1/1]
PD#SWPL-196422

Problem:
spsg works in trigger mode. When ldimming requests stop,
only the first part of the last frame of data is transmitted,
and the following data was not sent.

Solution:
When spisg is in the transmission state or idle state, the
DESC_DONE and DESC_LIST_DONE flags are both 1, which results
in incorrect judgment.
Change to use start.pending and descriptor.pending

Verify:
t3x bc302

Change-Id: Ib6b1b2f46e4f036bc2a5c8728afef67bbc2fca33
Signed-off-by: Sunny Luo <sunny.luo@amlogic.com>