uvm: uncompressed format doesn't need gpu realloc [2/3]
PD#SWPL-193049

Problem:
after modifying the nv12 format, the h264 process goes to
the ge2d copy, but the ge2d is not adapted to nv12, and
an error occurs during data copying.

Solution:
a compress member is added to decoder_para to determine
whether GPU reallocation is required.

Verify:
android-u

Change-Id: Ifa27e3eab0326f82244ed6e7bb986d69613107c5
Signed-off-by: yang.zhou <yang.zhou@amlogic.com>