hdmitx: Optimize the hdr_priority process on the Linux [1/1]
PD#SWPL-197215

Problem:
On Linux, after switching the UI to SDR and rebooting,
the output remains HLG. The reason is that the hdr_priority
did not clear the HLG status in hdr_info, causing the system
to still assume HLG is supported after reboot.

Solution:
When the hdr_priority value is 2, it is necessary to clear
the HDR10 and HLG status.

Verify:
yocto

Test:
DRM-TX-34

Change-Id: I2f4e09bfc0b5452f4b2ee90beec6835ee3d2b539
Signed-off-by: zhou.han <zhou.han@amlogic.com>