tee: optee: Allow to freeze the task waiting for tee-supplicant [1/1]
PD#SWPL-205216

Problem:
When the system is going to hibernate or suspend,
it might happen that the tee-supplicant task is frozen first.
wait_for_completion_interruptible might get stuck in this case.

Solution:
Add try_to_freeze to allow the waiting task to be frozen while
waiting for the response of tee-supplicant.

Verify:
T5D android U

Change-Id: Idcc4c25decb87f17db078992ba057fd592a21dba
Signed-off-by: Hangyu Li <hangyu.li@amlogic.com>