clk: Optimize the mechanism of the retry lock PLL [1/1]
PD#SWPL-202453

Problem:
The original retry mechanism was only effective when setting the PLL
rate. When enabling the PLL, there was a lack of a retry mechanism to
ensure PLL lock.

Solution:
In the function that enables the PLL, add a retry mechanism to lock the
PLL, ensuring that both when setting the PLL rate and enabling the PLL,
the PLL can be locked as much as possible.

Verify:
S905X4

Change-Id: I3b96941e2a433d03413c151e7d2e1ce8e8e67684
Signed-off-by: Chuan Liu <chuan.liu@amlogic.com>