uvm: uncompressed format doesn't need gpu realloc [2/2]
PD#SWPL-211950

Problem:
after modifying the nv12 format, the h264 process goes to
the ge2d copy, but the ge2d is not adapted to nv12, and
an error occurs during data copying.

Solution:
a compress member is added to decoder_para to determine
whether GPU reallocation is required.

Verify:
android-16

Change-Id: I640f5ceaf6fce3cf7b0146d8accc6fe26f6c9373
Signed-off-by: yang.zhou <yang.zhou@amlogic.com>