WiFi: fix the sdio device couldn't enum [1/1]
PD#SH-24116

Problem:
The power-on gpio is set to a high level by the wifi_dt driver,
then the power-on gpio is set to a low level by the bt_device driver.
This results in a pulse signal, cause the sdio device can not Enum.

Solution:
do not set the power-on gpio to a high level by wifi_dt driver.

Verify:
OK

Signed-off-by: hong.zhang <hong.zhang@amlogic.com>
Change-Id: I5cd0246512b85f1aeab37d7ed4ed364a81ee21d0
(cherry picked from commit f303874068296f32b440a921833cc13138f556c6)