hdmitx: optimise pfifo reset flow [1/1]
PD#SWPL-217845

Problem:
On a specific TV, rapidly switching modes may result in no signal
because resetting the TMDS gate can cause phase instability.

Solution:
Before setting the parameters of the PHY, reset the TMDS gate and
then check the stability of the phase. If the phase is unstable,
promptly reset the pfifo. When setting the PHY later, the TMDS gate
should not be reset again.

Verify:
s7/s7d/s6/s5

Change-Id: I58c514a7272f76fbe98121c56a6303fd1dd75619
Signed-off-by: zhou.han <zhou.han@amlogic.com>