hdmitx: uboot supports qms, kernel does not need to be configured [2/2]
PD#SWPL-228929

Problem:
the uboot mode is 59hz. When the device is powered on for
the first time and qms is enabled, the output under uboot
will switch to 60hz. After the kernel starts, the mode
passed by uboot to the kernel is 59hz, so qms is switched
to 59hz

Solution:
after uboot configures qms, kernel should not configure
qms again

Verify:
S905X5

Test:
DRM-TX-78

Change-Id: I264f9edd567863e0df1da1cbcf4533da825fac91
Signed-off-by: ruofei.zhao <ruofei.zhao@amlogic.com>