video: add multi plane for video display [2/2]
PD#SWPL-236654

Problem:
When DTV switches channels, a plane switching operation is performed.
When the video plane is disabled, the DRM releases the fence
corresponding to the current DMABUF.

Due to the differences in DMABUF usage by DTV, when the video
plane is re-enabled, it commits the DMABUF from before it was disabled.
Therefore, the DRM recognizes this as a repeat frame and does not
recreate the fence, continuing to use the fence from the previous frame.

Because the fence from the previous frame was already released when
the video plane was disabled, this results in a null pointer exception.

Solution:
When video is disabled, clear the DMABUF record to ensure that it is
not mistakenly identified as a repeat frame after being re-enabled.

Verify:
S7

Test:
DRM-OSD-20

Change-Id: Idb542a107adca7e90ec596959f412f8aef840c36
Signed-off-by: dijie.pan <dijie.pan@amlogic.com>