vdin: the hdr10 change was not reported with the HDR10 flag  [1/1]
PD#SWPL-242147

Problem:
Set HDR10, but no corresponding hdr10 flag was displayed.
Before vdin start dec, sdr->hdr was not set, so vdin did not report.

Solution:
As long as the signal is stable, even if start_dec is not completed,
the event will still be sent.

Verify:
t5m

Change-Id: Icffadda8e22eb9917217bef60d790fc6b2f506a3
Signed-off-by: mingxiu.sun <mingxiu.sun@amlogic.com>